requirements.txt
requests
beautifulsoup4
pandas
openpyxl
sentence-transformers
flask
gunicorn
Flask-APScheduler
python-docx
pdfplumber